+++ subprojects/mps-gradle-plugin/src/main/kotlin/com/specificlanguages/mps/MpsBuild.kt
@@ -72,7 +72,7 @@ abstract class MainBuild : MpsBuild() {
 
     /**
      * Indicates whether the build is internal or published (packaged). Published builds have their artifacts included
-     * in the package produced by `packageMps` and published by `components["mps"]`. By default, all main builds
+     * in the zip produced by `packageZip` and published by `components["mps"]`. By default, all main builds
      * are published.
      */
     abstract val published: Property<Boolean>